totalPrice

+++ src/app/Components/ShoppingCartitem/ShoppingCartitem.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
 import { ProductsService } from 'src/app/Services/products.service';
 import { OrderService} from 'src/app/Services/Order.service';
 
@@ -10,36 +10,45 @@ import { OrderService} from 'src/app/Services/Order.service';
 export class ShoppingCartitemComponent implements OnInit {
   quantity:any;
   product:any;
-  constructor(private productService:ProductsService,private orderService:OrderService) {
-
-  }
-
-  ngOnInit() {
-
-  }
+  OrderPrice:any;
   @Input() pd:any;
- 
-  increaseAmount(){
 
+  @Output() totalPriceOnChange: EventEmitter<number>;
+  
+  constructor(private productService:ProductsService,private orderService:OrderService) {
+   this.totalPriceOnChange = new EventEmitter<number>();
+  }
+  ngOnInit() {}
+                /**To be handled in services */
+  increaseAmount(){
     this.productService.GetProductById(this.pd.product.id).subscribe(
       (data)=>{
         this.product = data;
         this.quantity = this.product.quantity;
         if(this.pd.amount != this.quantity){
           this.pd.amount++;
+          this.OrderPrice = this.pd.product.price;
+          this.totalPriceOnChange.emit(this.OrderPrice);
           this.orderService.AddToShoppingCart(this.pd.product.id,localStorage.getItem('cartToken')).subscribe();
         }
       }
     );
-   
   }
-  removeItem(){
-    this.pd.amount--;
-    this.orderService.RemoveItemFromShoppingCart(this.pd.product.id,localStorage.getItem('cartToken')).subscribe();
-  }
-  removeItemTotalAmount(){
-    this.orderService.RemoveItemTotalAmountShoppingCart(this.pd.product.id,localStorage.getItem('cartToken')).subscribe();
-    window.location.reload();    
 
+  removeItem(){
+    if(this.pd.amount != 1){
+      this.pd.amount--;
+      this.OrderPrice = -1*this.pd.product.price;
+      this.totalPriceOnChange.emit(this.OrderPrice);
+      this.orderService.RemoveItemFromShoppingCart(this.pd.product.id,localStorage.getItem('cartToken')).subscribe();
+    }
+    if(this.pd.amount == 1)
+      this.removeItemTotalAmount()
   }
+
+  removeItemTotalAmount(){
+    this.orderService.removeItemTotalAmount(this.pd.product.id,localStorage.getItem('cartToken'));
+  }
+
+
 }
